Move `evaluate` to `Expression`
