+++ src/main/kotlin/sigma/semantics/expressions/Call.kt
@@ -1,5 +1,9 @@
 package sigma.semantics.expressions
 
+import sigma.evaluation.Thunk
+import sigma.evaluation.scope.Scope
+import sigma.evaluation.values.FunctionValue
+import sigma.evaluation.values.Value
 import sigma.semantics.Computation
 import sigma.semantics.TypeScope
 import sigma.semantics.DeclarationScope
@@ -129,5 +133,27 @@ class Call(
             subjectCallOutcome.value as? IllegalSubjectCallError,
             argumentValidationOutcome.value as? InvalidArgumentError,
         )
+    }
+
+    override fun evaluate(
+        scope: Scope,
+    ): Thunk = object : Thunk() {
+        override val toEvaluatedValue: Value
+            get() {
+                val subjectValue = subject.evaluate(scope = scope).toEvaluatedValue
+
+                if (subjectValue !is FunctionValue) throw IllegalStateException("Subject $subjectValue is not a function")
+
+                val argumentValue = argument.evaluate(scope = scope)
+
+                // Thought: Obtaining argument here might not be lazy enough
+                val image = subjectValue.apply(
+                    argument = argumentValue.toEvaluatedValue,
+                )
+
+                return image.toEvaluatedValue
+            }
+
+        override fun dump(): String = "(bound call)"
     }
 }
