+++ src/main/kotlin/sigma/semantics/expressions/FieldRead.kt
@@ -1,8 +1,11 @@
 package sigma.semantics.expressions
 
+import sigma.evaluation.Thunk
+import sigma.evaluation.scope.Scope
 import sigma.semantics.Computation
 import sigma.semantics.TypeScope
 import sigma.evaluation.values.Symbol
+import sigma.evaluation.values.tables.DictTable
 import sigma.semantics.DeclarationScope
 import sigma.semantics.SemanticError
 import sigma.semantics.types.IllType
@@ -114,5 +117,17 @@ class FieldRead(
             inferredSubjectTypeOutcome.value as? InvalidSubjectTypeError,
             inferredFieldTypeOutcome.value as? MissingFieldError,
         )
+    }
+
+    override fun evaluate(
+        scope: Scope,
+    ): Thunk {
+        val subjectValue = subject.evaluate(scope = scope).toEvaluatedValue
+
+        if (subjectValue !is DictTable) throw IllegalStateException("Subject $subjectValue is not a dict")
+
+        val value = subjectValue.apply(fieldName)
+
+        return value
     }
 }
