+++ src/main/kotlin/sigma/semantics/types/TupleType.kt
@@ -1,5 +1,7 @@
 package sigma.semantics.types
 
+import sigma.evaluation.scope.Scope
+import sigma.evaluation.values.tables.Table
 import sigma.semantics.SyntaxValueScope
 import sigma.semantics.expressions.Abstraction
 
@@ -11,4 +13,6 @@ abstract class TupleType : TableType() {
     abstract override fun substituteTypeVariables(
         resolution: TypeVariableResolution,
     ): TupleType
+
+    abstract fun toArgumentScope(argument: Table): Scope
 }
