+++ src/main/kotlin/sigma/syntax/typeExpressions/OrderedTupleTypeConstructorTerm.kt
@@ -46,12 +46,4 @@ data class OrderedTupleTypeConstructorTerm(
             )
         },
     )
-
-    override fun toArgumentScope(argument: Table): Scope = object : Scope {
-        override fun getValue(name: Symbol): Thunk? {
-            val index = elements.indexOfOrNull { it.name == name } ?: return null
-
-            return argument.read(IntValue(value = index.toLong()))
-        }
-    }
 }
