+++ src/test/kotlin/sigma/syntax/expressions/OrderedTupleConstructorTermTests.kt
@@ -79,50 +79,4 @@ class OrderedTupleConstructorTermTests {
             )
         }
     }
-
-    object EvaluationTests {
-        @Test
-        fun testEmpty() {
-            val value = ExpressionTerm.parse(
-                source = "[]",
-            ).evaluate(
-                scope = Scope.Empty,
-            ).toEvaluatedValue
-
-            assertIs<DictTable>(value)
-
-            assertEquals(
-                expected = ArrayTable(
-                    elements = emptyList(),
-                ).evaluatedEntries,
-                actual = value.evaluatedEntries,
-            )
-        }
-
-        @Test
-        fun testNonEmpty() {
-            val value = ExpressionTerm.parse(
-                source = "[a, b]",
-            ).evaluate(
-                scope = FixedScope(
-                    entries = mapOf(
-                        Symbol.of("a") to BoolValue(false),
-                        Symbol.of("b") to IntValue(1),
-                    ),
-                ),
-            ).toEvaluatedValue
-
-            assertIs<DictTable>(value)
-
-            assertEquals(
-                expected = ArrayTable(
-                    elements = listOf(
-                        BoolValue(false),
-                        IntValue(1),
-                    ),
-                ).evaluatedEntries,
-                actual = value.evaluatedEntries,
-            )
-        }
-    }
 }
